requirements.txt
streamlit
langchain
faiss-cpu
sentence-transformers
beautifulsoup4
requests
